+++ src/App.jsx
@@ -8,9 +8,9 @@ import BrowseProjects from './pages/BrowseProjects';
 import ProjectDetails from './pages/ProjectDetails';
 import CodeCrewLayout from './layouts/CodeCrewLayout';
 import ForgotPassword from './pages/ForgotPassword';
-import Profile from './pages/Profile';
+import EditProfile from './pages/EditProfile';
 import { UserProvider } from './components/UserContext';
-import UserProfile from './pages/UserProfile';
+import Profile from './pages/Profile';
 
 
 const URL = 'http://localhost:8000/api/v1/';
@@ -42,9 +42,9 @@ const App = () => {
           <Route path="/projects" element={<BrowseProjects />} />
           <Route path="/projects/:id" element={<ProjectDetails />} />
           <Route path="/forgot-password" element={<ForgotPassword />} />
-          <Route path="/profile" element={<Profile />} />
+          <Route path="/edit-profile" element={<EditProfile />} />
           <Route path="/browse-projects" element={<BrowseProjects />} />
-          <Route path="/my-profile" element={<UserProfile/>} />
+          <Route path="/my-profile" element={<Profile />} />
         </Routes>
       </CodeCrewLayout>
     </UserProvider>
